jsp 파일 업로드 공부 끝

+++ File_Upload/WebContent/uploadAction.jsp
@@ -1,5 +1,6 @@
 <%@ page import="model.dao.FileDAO" %>
 <%@ page import="java.io.File" %>
+<%@ page import="java.util.Enumeration" %>
 <!-- 사용자가 업로드한 파일 중 이름 똑같으면 자동으로 파일 이름 바꿔주고 오류 발생하지 않게 해주는 클래스 -->
 <%@ page import="com.oreilly.servlet.multipart.DefaultFileRenamePolicy" %>
 <!-- 파일 업로드 수행하는 클래스 -->
@@ -16,7 +17,7 @@
 	<%
 		//String directory = application.getRealPath("/upload/");
 		// 하드코딩으로 업로드 경로 루트 디렉터리 밖으로 이동
-		String directory = "D:/git/JSP_Study_dongbina/upload";
+		String directory = "D:/git/JSP_Study_dongbina/upload/";
 		int maxSize = 1024 * 1024 * 100; // 최대 100MB만 저장 가능하게
 		String encoding = "UTF-8";
 		
@@ -25,22 +26,32 @@
 			= new MultipartRequest(request, directory, maxSize, encoding,
 					new DefaultFileRenamePolicy());
 		
-		String fileName = multipartRequest.getOriginalFileName("file");
-		String fileRealName = multipartRequest.getFilesystemName("file");
+		// for문과 같은 형식으로 사용
+		Enumeration fileNames = multipartRequest.getFileNames();
 		
-		// secure coding 1 업로드 가능한 확장자 명확히 제시하기
-		// 해당 확장자 아니면 파일 지워버림
-		if (!fileName.endsWith(".doc") && !fileName.endsWith(".hwp")
-				&& !fileName.endsWith(".pdf") && !fileName.endsWith(".xls")) {
-			File file = new File(directory + fileRealName);
-			file.delete();
-			out.write("업로드 할 수 없는 확장자입니다.");
-		} else {
-			new FileDAO().upload(fileName, fileRealName);
+		while (fileNames.hasMoreElements()) {
+			String parameter = (String) fileNames.nextElement();
 			
-			out.write("파일명 : " + fileName + "<br>");
-			out.write("실제 파일명 : " + fileName + "<br>");
+			String fileName = multipartRequest.getOriginalFileName(parameter);
+			String fileRealName = multipartRequest.getFilesystemName(parameter);
+			
+			if (fileName == null) continue;
+			
+			// secure coding 1 업로드 가능한 확장자 명확히 제시하기
+			// 해당 확장자 아니면 파일 지워버림
+			if (!fileName.endsWith(".doc") && !fileName.endsWith(".hwp")
+					&& !fileName.endsWith(".pdf") && !fileName.endsWith(".xls")) {
+				File file = new File(directory + fileRealName);
+				file.delete();
+				out.write("업로드 할 수 없는 확장자입니다." + "<br>");
+			} else {
+				new FileDAO().upload(fileName, fileRealName);
+				
+				out.write("파일명 : " + fileName + "<br>");
+				out.write("실제 파일명 : " + fileName + "<br>");
+			}
 		}
+		
 	%>
 </body>
 </html>
